requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
streamlit==1.10.0
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.6
matplotlib==3.5.1
seaborn==0.11.2
requests==2.27.1
beautifulsoup4==4.11.1
schedule==1.1.0
python-dotenv==0.19.2
slack-sdk==3.19.5
sqlalchemy==2.0.23
mysqlclient==2.1.1
redis==3.5.3
pydantic==2.5.2
python-multipart==0.0.6
boto3==1.20.32
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-env==1.0.1
pytest-xdist==3.3.1
selenium==4.16.0 